requirements.txt
google-generativeai
python-dotenv
flask
scikit-learn
pandas 
numpy
matplotlib